add layout for data-binding
