--- app/src/main/java/com/fazahidah/dodolan/LoginActivity.kt
@@ -1,20 +1,0 @@
-package com.fazahidah.dodolan
-
-import android.content.Intent
-import android.os.Bundle
-import android.widget.Button
-import androidx.appcompat.app.AppCompatActivity
-
-class LoginActivity: AppCompatActivity() {
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.login_page)
-
-        val login = findViewById<Button>(R.id.button_login)
-        login.setOnClickListener {
-            val intent = Intent(this, MainActivity::class.java)
-            startActivity(intent)
-        }
-    }
-}
