+++ app/src/main/java/com/fazahidah/dodolan/ui/activity/LoginActivity.kt
@@ -1,0 +1,29 @@
+package com.fazahidah.dodolan.ui.activity
+
+import android.content.Intent
+import android.os.Bundle
+import androidx.appcompat.app.AppCompatActivity
+import androidx.databinding.DataBindingUtil
+import com.fazahidah.dodolan.R
+import com.fazahidah.dodolan.databinding.LoginBinding
+
+class LoginActivity: AppCompatActivity() {
+
+    private lateinit var binding: LoginBinding
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        binding = DataBindingUtil.setContentView(this, R.layout.login)
+
+        // binding actions
+        binding.buttonLogin.setOnClickListener {
+            val intent = Intent(this, MainActivity::class.java)
+            startActivity(intent)
+        }
+
+        binding.register.setOnClickListener {
+            val intent = Intent(this, RegisterActivity::class.java)
+            startActivity(intent)
+        }
+    }
+}
