+++ app/src/main/java/com/fazahidah/dodolan/utils/Utils.kt
@@ -1,0 +1,7 @@
+package com.fazahidah.dodolan.utils
+
+import android.app.Activity
+import android.content.Context
+import android.content.Intent
+import androidx.core.content.ContextCompat.startActivity
+
