+++ app/src/main/res/layout/login.xml
@@ -2,43 +2,36 @@
 <layout>
 
     <androidx.constraintlayout.widget.ConstraintLayout
+        xmlns:tools="http://schemas.android.com/tools"
         xmlns:android="http://schemas.android.com/apk/res/android"
         xmlns:app="http://schemas.android.com/apk/res-auto"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         android:background="@drawable/bg">
 
-        <androidx.cardview.widget.CardView
-            android:id="@+id/cardView"
+        <Button
+            android:id="@+id/button_login"
             android:layout_width="170dp"
             android:layout_height="50dp"
-            android:layout_marginBottom="184dp"
-            app:cardBackgroundColor="#094776"
-            app:cardCornerRadius="50dp"
+            android:layout_marginBottom="216dp"
+            android:text="@string/login"
+            android:background="@drawable/mybutton"
+            android:textColor="#FFFFFF"
             app:layout_constraintBottom_toBottomOf="parent"
             app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintHorizontal_bias="0.497"
+            app:layout_constraintHorizontal_bias="0.498"
             app:layout_constraintStart_toStartOf="parent">
 
-            <TextView
-                android:id="@+id/textView"
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:layout_gravity="center"
-                android:text="@string/registerbutton"
-                android:textColor="@android:color/background_light"
-                android:textSize="24sp"
-                app:fontFamily="sans-serif" />
-        </androidx.cardview.widget.CardView>
+        </Button>
 
         <androidx.cardview.widget.CardView
             android:id="@+id/cardView2"
             android:layout_width="280dp"
-            android:layout_height="300dp"
+            android:layout_height="180dp"
             android:layout_marginTop="8dp"
             android:layout_marginBottom="15dp"
             app:cardCornerRadius="15dp"
-            app:layout_constraintBottom_toTopOf="@+id/cardView"
+            app:layout_constraintBottom_toTopOf="@+id/button_login"
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintHorizontal_bias="0.496"
             app:layout_constraintStart_toStartOf="parent"
@@ -51,7 +44,7 @@
                 android:layout_height="wrap_content"
                 android:layout_gravity="center_horizontal"
                 android:layout_marginTop="40dp"
-                android:autofillHints="@string/username_email"
+                android:autofillHints="@string/app_name"
                 android:ems="10"
                 android:hint="@string/username_email"
                 android:inputType="textEmailAddress" />
@@ -74,14 +67,14 @@
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:layout_marginStart="77dp"
-            android:text="@string/has_acc"
+            android:text="@string/no_acc"
             android:textColor="#ffff"
             android:textSize="8pt"
             app:layout_constraintBottom_toBottomOf="parent"
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintHorizontal_bias="0.045"
             app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toBottomOf="@+id/cardView"
+            app:layout_constraintTop_toBottomOf="@+id/button_login"
             app:layout_constraintVertical_bias="0.163" />
 
         <TextView
@@ -89,12 +82,26 @@
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:layout_marginEnd="77dp"
-            android:text="@string/login"
+            android:text="@string/register"
             android:textColor="#094776"
             android:textSize="8pt"
-            app:layout_constraintStart_toEndOf="@id/reg"
-            app:layout_constraintTop_toTopOf="@id/reg"
-            app:layout_constraintEnd_toEndOf="parent"/>
+            app:layout_constraintBottom_toBottomOf="parent"
+            app:layout_constraintEnd_toEndOf="parent"
+            app:layout_constraintHorizontal_bias="1.0"
+            app:layout_constraintStart_toStartOf="parent"
+            app:layout_constraintTop_toBottomOf="@+id/button_login"
+            app:layout_constraintVertical_bias="0.163" />
+
+        <ImageView
+            android:id="@+id/imageView"
+            android:layout_width="300dp"
+            android:layout_height="137dp"
+            app:layout_constraintBottom_toTopOf="@+id/cardView2"
+            app:layout_constraintEnd_toEndOf="parent"
+            app:layout_constraintStart_toStartOf="parent"
+            app:layout_constraintTop_toTopOf="parent"
+            app:srcCompat="@drawable/logo"
+            tools:ignore="ContentDescription" />
 
 
     </androidx.constraintlayout.widget.ConstraintLayout>
